ngteste/src/app/shared/services/unique-id/unique-id.service.spec.ts
import {UniqueIdService} from './unique-id.service';

describe(UniqueIdService.name, () => {

  let service: UniqueIdService = null;

  beforeEach(() => {
    service = new UniqueIdService();
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name} should generate id when called whith prefix`, () => {
    const id = service.generateUniqueIdWithPrefix('app');
    expect(id.startsWith('app-')).toBeTrue();
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name} should not genarate duplicate IDs when called multiple times`, () => {
    const ids = new Set();
    for (let i = 0; i < 50; i++) {
      ids.add(service.generateUniqueIdWithPrefix('app'));
    }
    expect(ids.size).toBe(50);
  });

  it(`#${UniqueIdService.prototype.getNumberOfGneratedUniqueId.name} should return the number of genarate when called`, () => {
    service.generateUniqueIdWithPrefix('app');
    service.generateUniqueIdWithPrefix('app');
    expect(service.getNumberOfGneratedUniqueId()).toBe(2);
  });

  it(`#${UniqueIdService.prototype.generateUniqueIdWithPrefix.name} should throw when called empty `, () => {
    const emptyValues = [null, undefined, '', '0', '1'];
    emptyValues.forEach(emptyValue => {
      expect(() => service.generateUniqueIdWithPrefix(emptyValue))
        .withContext(`Empty value: ${emptyValue}`)
        .toThrow();
    })
  });
});

//describe(HomeComponent.name, () => {
//   it('#${HomeComponent.prototype.consultaGravame.name} should consultar gravame na cip when preenchido os campos cliente, contrato, requisição e dataInicial/dataFinal', () => {
//     const component = new HomeComponent();
//     const consulta =  component.consultaGravame('não sei ainda');
//     expect(id).toContain('não sei ainda')
//     });
// });
